+++ scripts/imitation_learning/robomimic/play.py
@@ -72,6 +72,7 @@ import torch
 if args_cli.enable_pinocchio:
     import isaaclab_tasks.manager_based.locomanipulation.pick_place  # noqa: F401
     import isaaclab_tasks.manager_based.manipulation.pick_place  # noqa: F401
+    import isaaclab_tasks.manager_based.manipulation.pour_water  # noqa: F401
 
 from isaaclab_tasks.utils import parse_env_cfg