+++ source/isaaclab_mimic/isaaclab_mimic/envs/pinocchio_envs/__init__.py
@@ -35,6 +35,13 @@ gym.register(
 )
 
 gym.register(
+    id="Isaac-PourWater-GR1T2-Pink-IK-Abs-Mimic-v0",
+    entry_point=f"{__name__}.pour_water_gr1t2_mimic_env:PourWaterGR1T2MimicEnv",
+    kwargs={"env_cfg_entry_point": f"{__name__}.pour_water_gr1t2_mimic_env_cfg:PourWaterGR1T2MimicEnvCfg"},
+    disable_env_checker=True,
+)
+
+gym.register(
     id="Isaac-ExhaustPipe-GR1T2-Pink-IK-Abs-Mimic-v0",
     entry_point=f"{__name__}.pickplace_gr1t2_mimic_env:PickPlaceGR1T2MimicEnv",
     kwargs={"env_cfg_entry_point": f"{__name__}.exhaustpipe_gr1t2_mimic_env_cfg:ExhaustPipeGR1T2MimicEnvCfg"},